+++ CopilotKit/packages/backend/src/lib/openai-adapter.ts
@@ -40,15 +40,27 @@ export class OpenAIAdapter implements CopilotKitServiceAdapter {
       maxTokensForOpenAIModel(forwardedProps.model || this.model),
     );
 
-    const stream = this.openai.beta.chat.completions
-      .stream({
+    return new Promise((resolve, reject) => {
+      // remove message.function_call.scope if it's present.
+      // scope is a field we inject as a temporary workaround (see elsewhere), which openai doesn't understand
+      messages.forEach((message) => {
+        if (message.function_call?.scope) {
+          delete message.function_call.scope;
+        }
+      });
+
+      const stream = this.openai.beta.chat.completions.stream({
         model: this.model,
         ...forwardedProps,
         stream: true,
         messages: messages as any,
-      })
-      .toReadableStream();
-
-    return { stream };
+      });
+      stream.on("error", (error) => {
+        reject(error); // Reject the promise with the error
+      });
+      stream.on("connect", () => {
+        resolve({ stream: stream.toReadableStream() });
+      });
+    });
   }
 }