+++ CopilotKit/packages/react-core/src/hooks/index.ts
@@ -7,3 +7,4 @@ export { useCopilotAction } from "./use-copilot-action";
 export { useMakeCopilotReadable } from "./use-make-copilot-readable";
 export { useMakeCopilotDocumentReadable } from "./use-make-copilot-document-readable";
 export { type UseChatHelpers } from "./use-chat";
+export { useCopilotReadable } from "./use-copilot-readable";